This was a *major* overhaul in terms of code organization.
The codebase is much better organized and flexible now. In addition, it is the beginning of the switch to much greater configurability. Many hard-coded assumptions are gone, and the rest are in a much better position to use configuration options.

There are a few features in addition to the major cleanup. New commands for getting started easier: 'init' and 'new'. Also much better help text. Finally, the most significantly useful feature is that the refresh behavior is much better when using the design mode. Now, the data and templates update each time you refresh in addition to just the CSS. The next step is to add watch behavior to automatically live-reload the page when files change.